+++ src/is-text-selected.js
@@ -1,0 +1,13 @@
+import getSelectionElements from './get-selection-overlay-elements';
+
+
+/**
+ * @returns {boolean}
+ * Text selection is exists (at least one line).
+ */
+export default function isTextSelected() {
+    const selectionElements = getSelectionElements();
+    const isSelected = selectionElements.some((i) => !!i);
+
+    return isSelected;
+}